Add LogCommands option to log commands requested by client
At some point in the past, all logging of commands requested by the client
was removed from the OpenSSH server code with the rationale that this
preserves privacy.  Unfortunately, this preserves privacy at the cost of
safety and accountability - or, through the classical lens, enhances
user confidentiality at the expense of system availability and integrity.
Add a new LogCommands option - disabled by default! - to turn logging of
the full command requested by the client back on.